fix: ensure quit is called on outlook.application
Closes #3

This commit ensures that `Destroy` calls `Quit` on `mMailApp`
rather than just releasing the reference.

Error handling to output errors to the console is added to accomodate
this.

The Class_Terminate event is also added to ensure `Destroy` is called
automatically when the class goes out of scope.

The method signature for `GenerateEmail` is also reformatted for
readability and the `ByVal` keyword added to the optional parameters.

Finally some description annotations are added to make the class
attributes manageble via Rubberduck.