Minor cleanup and additions
Function names and headers modified for brevity. Removed manual enumeration for built-in enums.
Added 2 new functions into the ExcelAPIs file.

GetHeader - Returns the column number of a given header, from a given row of a given worksheet.

GetHeaders - Returns a dictionary object with the same concept as the GetHeader function. Dictionary key is the header text, and the item value is the header's column number.

These two functions are great when working with documents that might not have static column ordering. Or when you want to keep your code completely dynamic. For instance: Cells(RowNum, GetHeader(ActiveSheet, 1, "First Name"))